backend/requirements.txt
django
djangorestframework
psycopg2-binary
django-filter
gunicorn
django-cors-headers
pytz